bump to 0.2.0
Bump from 0.1.0 to 0.2.0 because there is a backward incompatible change:

- `label.padding` is renamed to `box.padding`

In `geom_label_repel()`, we now have `label.padding` and `box.padding`. The
`box.padding` parameter is used to add extra space around the bounding boxes
for computing overlap. The `label.padding` parameter controls the size of the
rectangular label around the text.