add nudge_x and nudge_y
This commit adds nudge_x and nudge_y to control the starting position of text
labels. The labels repel from each other and repel from data points. However,
there is a spring force that attracts each label to its own data point. You
can use nudge_x and nudge_y to adjust the starting position of the label and
the origin of the spring force.