Make the segment the same color as the text
Instead of setting segment.color to a mid-level grey, make it a bit
transparent and color it according to the color of the text. By default
the text is black and the half-transparent segment therefore still
appears grey.

When the text itself is made transparent, the transparency of the text
and of the segment are combined (multiplied).

Another possibility would be to declare two new *mappable* aesthetics:
segment.color and segment.alpha, and let the user do the mapping (i.e.
geom_text_repel(color=x, segment.color=x)) which would allow to color
the text and segment separately. But I find it quite cumbersome for a
small gain in versatility.