fix #115 heap-buffer-overflow
Running ASAN revealed that I had a heap buffer overflow bug in my C++
code:

    SUMMARY: AddressSanitizer: heap-buffer-overflow /data/gannet/ripley/R/packages/tests-clang-SAN/ggrepel/src/repel_boxes.cpp:631:25 in repel_boxes

Basically, the problem goes like this:

1. Allocate a vector with N elements.
2. Try to access the (N+1) element --> heap buffer overflow error

I hope this fixes the issue!

This crashed R on Windows but not on macOS or Linux. Isn't that strange?
I don't understand why the crash depends on the operating system.