Add support for grouping and facets to nudge functions
As position_nudge_repel() and position_nudge_repel_t() compute summaries from data or fit splines or polynomials, there is a need to obey grouping and facets. As the grouping may be best ignored in some cases, I replaced the compute_layer function by a compute_panel function and included in this function the code to support grouping.
There was also a problem in position_nudge_repel_t() with the computation of the angles, which required taking into consideration the range of the data, for this to work consistently across groups, we get the information from scales, which is a parameter of the panel function, but not of the layer function.
For what I have tested, things seem to work reasonably well. 
I haven't thoroughly tested if the use of the compute_panel function changes the behaviour compared to position_nudge(). It also remains a test of facets with free scales. I would expect this not to work well, but I do not know if it works with position_nudge().