Added three new aesthetics:
- label.col.alpha
- label.fill.alpha

these can be used to decouple transparency values of label background and frame from the overall transparency of the shape, thus enabling the text to be rendered without alpha while having the possiblity to make the background transparent.

if any of the two (or both) are defined (not NULL), their values will override the general alpha parameter, and if not, alpha will be used to make everything transparent.

- segment.linewidth

segment.size is somewhat misleading as it implies the length of the segment, so segment.linewidth can be used instead (it will not be converted to points like segment.size, though)